automation/pyproject.toml
[project]
name = "automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["silverback==0.7.15", "ape-alchemy==0.8.8"]

[tool.uv]
dev-dependencies = ["ruff>=0.11.12"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "F", # pyflakes
    "I", # isort
]
